+++ src/main/java/subway/domain/PathMenu.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.domain;
 
 import java.util.Arrays;
